build/xilinx/ise: Use separate TNM_NET for period and TIG constraints.
Because ISE is retarded, if you have anything other than period constraints on
a TNM_NET (such as a TIG), then it is unable to trace the constraints through
any of the clock elements like PLLs and DCMs. It will generate the following
warning instead;

> ConstraintSystem - TNM : *** was distributed to a DCM but new TNM constraints
> were not derived. The requirement for derived TNM constraints is that the
> distributed TNM is referenced by no more than a single PERIOD constraint.
> Non-PERIOD referencers are also not allowed.

The requirements for the TNM to be derived is described at
[Timing Constraints User Guide - UG612 (v 11.1.1) April 29, 2009](http://www.xilinx.com/support/documentation/sw_manuals/xilinx11/ug612.pdf):

 * The TNM group is used in exactly one PERIOD specification.
 * The TNM group is not used in any FROM-TO or OFFSET specifications.
 * The TNM group is not referenced in any user group definition.

The fix is described in http://www.xilinx.com/support/answers/40007.html ||
http://www.xilinx.com/support/answers/37782.html

Create two TNM Groups for your input clock. One used for the FROM-TO
constraints and the other for the PERIOD propagation.

Fixes https://github.com/m-labs/migen/issues/43