altera.programmer: fix Windows-ism in load_bitstream.
To specify a particular programmer, one has to specify it as e.g.
"USB-Blaster [USB-0]" on Windows (where 0 is its number on
the system) or e.g. "USB-Blaster [2-1.2]" on Linux (where 2-1.2
is its USB path). The existing default therefore did not work on
Linux.

This commit fixes the default to pick the first USB-Blaster, and
renames the argument to deliberately break old code that relies
on this nonportable behavior.