cdc: make pipelined registers reset_less
data in pipelined registers is guaranteed to be clocked in from source
signals at "full rate" and fully determined by those source registers.
in that sense these registers don't hold intrinsic state that needs to
be reset explicitly. their data can just flow in from the source (which
is reset).
having them reset_less delays their reset by one (or a few) clock cycles.
if the target CD is reset, the data would just be reset for the duration
of the target CD reset and then ("slowly") transition back to what the
source CD has.
control and handshaking logic needs to be reset.