fhdl/verilog: list available clock domains in case of unresolved clock domain
Migen automatically renames some clock domains (for example in the case of
multiple modules defining a clock domain with the same name). When designing,
we need in some cases to know the final name of the clock domain and displaying
the list of avalaible clock domains helps figuring out what it is.

We are using Exception instead of KeyError since KeyError is not able to display
on multiple lines.